refactor: extract PDF list marker handling
